Title: 서로 다른 숫자, Time: 91ms, Memory: 2MB, Status: Passed - CodeTree

+++ 231207/서로 다른 숫자/distinct-numbers.cpp
@@ -4,6 +4,7 @@
 
 using namespace std;
 
+// 변수 선언
 int n;
 unordered_set<int> s;
 
@@ -12,8 +13,11 @@ int main() {
     for(int i=0; i<n; i++){
         int val;
         cin >> val;
+        // HashSet에 원소를 추가
         s.insert(val);
     }
-    cout << s.size();
+
+    // HashSet의 크기를 출력
+    cout << (int)s.size();
     return 0;
 }
